Updated renderer to handle multi-element card output
refs https://github.com/TryGhost/Team/issues/2447

Some cards like `markdown` return multiple elements instead of a single parent element as the output of `exportDOM`, so fetching `outerHTML` element for them won't be correct. Instead, these cards will return the elements as `innerHTML` of a div so our lexical renderer can fetch the inner HTML and render that as a result. This change -

- handles a new `type` property returned by `exportDOM` which can have option as `inner` or `outer`. `inner` type elements are handled by fetching their `innerHTML` data.