Extracted Unsplash-Selector Component package (#1165)
no issue

We currently have 3 versions of the Unsplash selector; The Admin-X version, written in Typescript + React; The Koenig version, written in Javascript + React; and the original Ember version. 
The react versions, are all implemented the same way, which leaves us with duplicate code of the same thing.

This creates a standalone Unsplash-Selector package, extracted from the Admin-X version.

The plan is to only have 1 package to maintain and be able to reuse it elsewhere to avoid having to maintain separate versions of the selector.