🐛 Fixed snippets displaying for users without permissions (#1478)
ref https://linear.app/ghost/issue/ONC-815

We were always displaying the snippet buttons (toolbar + delete button) despite the underlying functions not always being present, resulting in broken buttons. We control the ability to create/delete snippets via passing the fns, so if the fns don't exist we simply shouldn't display the buttons.

This does not include tests for the floating toolbar button as we don't currently have support for custom cardConfig in those tests (with the demo app) and adding this was not cooperative.